+++ src/main/java/racing/domain/CarRacingProperty.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public class CarRacingProperty {
