+++ src/main/java/racing/domain/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public interface MoveStrategy {
     boolean movable();
